requirements.txt
langchain==0.3.14
langchain-openai==0.3.1
langchain-community==0.3.14
openai==1.63.2
PyGithub==2.5.0
click==8.1.8
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
pydantic==2.10.6
tenacity==9.0.0
colorlog==6.9.0
python-jenkins==1.8.0
python-gitlab==5.4.0
tiktoken==0.9.0
# ── Web UI ───────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
